Fixed a few bugs.
First was an edge case for the approximation, which is when the function
is given 2 or fewer points as input. This is a base case, so it just
returns. The other was an incorrect indexing for partitioning, since
Python's array slicing is not inclusive. Last, a few changes were made
to comments.