Fixed a few bugs with the polygonal approximation.
The smaller bugs were an incorrect assertion, and incorrect arguments to
several numpy functions, which expect the inputs as a list. The larger
bug was the incorrect specification of the line vector, which had an
incorrect sign for the 'c' value, and needed to be normalized by the
slope's norm.

Also, switched to using `concatenate` instead of `hstack` and `vstack`,
as they will be deprecated.