Fixed an issue with the regression line computed.
The problem was that the code was unable to handle vertical or near
vertical regression lines. The issue stems from the fact that there is
not real correlation between the x and y points of the polyline, as they
define an arbitrary curve. Least squares regression
(`numpy.linalg.lstsq`) assumes that the y points are a function of the x
points, and it therefore cannot generate a vertical regression line.

The solution is to instead use total linear least squares (AKA
orthogonal distance) regression. This seeks to minimize the orthogonal
distance from the regression line to the points, and can be easily
implemented with principal component analysis (PCA) decomposition.