Fixed a few major issues with the implementation.
The first issue was that, when the points were partitioned. The line
segment between the two points was being dropped. The paper has it this
way, but doing so doesn't make sense because this is removing part of
the curve. Now, the split happens "at" a point, so the two new subsets
both contain the point.

The second major issue was the line being used to split the points. In
the paper, it says to use a "regression line" based on the points.
However, using a true regression line doesn't make sense because the
algorithm should be looking at the thickness of the polyline if the
given section was reduced to the line segment between the endpoints.
Since the regression line will not, in general, pass the through the
endpoints, this will not be the case. Thus, we now instead use the line
between the endpoints.