Added in handling of duplicate points.
Since the splitting of points is now inclusive, the same point can end
up in two sub-polylines of the larger polyline, and therefore can be
added to the dominant point list twice. This is handled by using an
ordered dictionary, which makes the points unique, and preserves the
original ordering, which is important for preserving the polyline.