Made the unit test independent of where it is run.
When adding the `src` directory to the path, the earlier code relied on
the fact that the script was run inside the `unit_tests` directory. In
order to make this more general, the internal python variable `__file__`
is used, which contains the path to the current file being executed.

Also, improved the formatting for the printed results for one of the
tests.