Added a script to run polygonal approximation on a directory of images.
This script expects the images to be in a format similar to the of the
MPEG-7 database, which is a binary image that contains a single, closed
shape. The script uses OpenCV to get an initial contour, then runs our
approximation to get the approximated one, and saves these overlaid on
the image to file.