Changed the background subtraction interface to process single frames.
This is as opposed to before, where it attempted to process all
available frames. For a video file this is fine, but this will not work
well for a camera device. Thus, the interface for the functions were
switched to streaming based pattern, where the function now processes
the next video frame from the stream.

The camera and video functions were also combined, so the single
interface function now treats all streams the same. Some general cleanup
happened to the code as well, and the thickedge application was updated
appropriately.