Enable camera to be specified as inputs.
The thickedge application now allows for cameras to be specified as the
input. The application will then overlay polygons on the foreground
objects in real time.

Also, the handling of keys by the application has been improved. The
application will now terminate if the user presses 'q', or if a keyboard
interrupt is received.