Added in the ability to reset the BGS algorithm's state.
When the user presses 'r', the program re-initializes the BGS algorithm,
clearing its state. This is useful for some algorithms which are very
stateful, and can occasionally have errors that get stuck in their
output.

Also, the BGS algorithm is now passed as a parameter to functions,
instead of being a global variable.